backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
supabase>=2.3.0
pandas>=2.2.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
# Database/Async
asyncpg>=0.29.0
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
# AI / LLM
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.10.0
tiktoken>=0.5.0
